refactor(test-util): move PartitionsActuatorClient to test-util
Moves the PartitionsActuatorClient from the update-tests module to the
more general test-util module. This way any integration tests which
wants to query the actuators for information can use the client.
