+++ qa/update-tests/src/test/java/io/camunda/zeebe/test/ContainerState.java
@@ -8,6 +8,7 @@
 package io.camunda.zeebe.test;
 
 import io.camunda.zeebe.client.ZeebeClient;
+import io.camunda.zeebe.test.util.actuator.PartitionsActuatorClient;
 import io.camunda.zeebe.test.util.testcontainers.RemoteDebugger;
 import io.camunda.zeebe.test.util.testcontainers.ZeebeTestContainerDefaults;
 import io.zeebe.containers.ZeebeContainer;
