+++ qa/update-tests/src/test/java/io/camunda/zeebe/test/RollingUpdateTest.java
@@ -14,7 +14,8 @@ import io.camunda.zeebe.client.api.response.ProcessInstanceEvent;
 import io.camunda.zeebe.client.api.worker.JobHandler;
 import io.camunda.zeebe.model.bpmn.Bpmn;
 import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
-import io.camunda.zeebe.test.PartitionsActuatorClient.PartitionStatus;
+import io.camunda.zeebe.test.util.actuator.PartitionsActuatorClient;
+import io.camunda.zeebe.test.util.actuator.PartitionsActuatorClient.PartitionStatus;
 import io.camunda.zeebe.test.util.asserts.EitherAssert;
 import io.camunda.zeebe.test.util.asserts.TopologyAssert;
 import io.camunda.zeebe.test.util.testcontainers.ZeebeTestContainerDefaults;
